Fixed The conclusion of this test is that the process runtime mode of qtchart-based GraphDialog is at present too slow

+++ GUI/charts/GraphDialogN.h
@@ -105,7 +105,7 @@ public:
     gstring mainModuleName() const
     {  return  pModule->GetName();   }
 
-    void AddPoint( int nPlot, int nPoint );
+    void AddPoint( size_t nPlot, int nPoint );
     void ShowGraph( const char * capAdd = nullptr );
 
 private:
